+++ RELEASE_NOTES.md
@@ -2,6 +2,7 @@
 
 ### 0.6.2  (WIP)
 - #121 Adds command to Open Root Directory
+- #128 Adds command to Open Default Path
 - #120 Adds command to Repeat Previous Search
 - #116 Adds ability to delete search history
 - #114 Adds "How To Use Koffee" link in settings window